pyproject.toml
[tool.poetry]
name = "sqltrace"
version = "0.1.0"
description = ""
authors = ["Adrian Garcia Badaracco <1755071+adriangb@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
google-cloud-sqlcommenter = { version = ">=2", extras = ["opentelemetry"]}
opentelemetry-api = ">=1"
typing-extensions = ">=4"
opentelemetry-sdk = "^1.23.0"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
psycopg = "^3.1.18"
psycopg-pool = "^3.2.1"
opentelemetry-sdk = "^1.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
